DL/SFESmoke/requirements.txt
wheel
fire
h5py
Keras==2.1.5
numpy
PyQt5==5.10.1
requests
tensorflow-gpu==1.4.0
cherrypy
pycrypto
pymysql
sqlalchemy